+++ src/EasyRanges.jl
@@ -11,7 +11,10 @@ using Base: OneTo
 
 rewrites range expression `expr` with extended syntax. The result is an `Int`-valued index
 range (possibly Cartesian) where indices are running in the forward direction (with a
-positive step).
+positive step). Call [`@reverse_range`](@ref) if a negative step is required.
+
+See [`EasyRanges.normalize`](@ref) to implement non-standard index or range types in the
+`@range` and [`@reverse_range`](@ref) macros.
 
 """
 macro range(ex::Expr)
@@ -23,7 +26,10 @@ end
 
 rewrites range expression `expr` with extended syntax. The result is an `Int`-valued index
 range (possibly Cartesian) where indices are running in the reverse direction (with a
-negative step).
+negative step). Call [`@range`](@ref) if a positive step is required.
+
+See [`EasyRanges.normalize`](@ref) to implement non-standard index or range types in the
+[`@range`](@ref) and `@reverse_range` macros.
 
 """
 macro reverse_range(ex::Expr)
@@ -53,21 +59,64 @@ function rewrite!(ex::Expr)
 end
 
 """
+    EasyRanges.normalize(x)
+
+yields an object which represent the same (Cartesian) index or set of indices as `x` but
+in one of the following forms:
+
+- an integer `i::Int` if `x` is equivalent to a single linear index;
+
+- a range of integers `r::AbstractRange{Int}` if `x` is equivalent to a range of linear
+  indices;
+
+- a multi-dimensional Cartesian index `I::CartesianIndex{N}` if `x` is equivalent to a
+  single `N`-dimensional Cartesian index;
+
+- a multi-dimensional Cartesian range `R::CartesianIndices{N}` if `x` is equivalent to a
+  rectangular region of `N`-dimensional Cartesian indices.
+
+This method may be extended by foreign packages to let `EasyRanges` known how to deal with
+other types of indices or of set of indices provided they are equivalent to one of the
+above canonical forms.
+
+"""
+normalize(x::Int) = x
+normalize(x::Integer) = Int(x)
+normalize(x::Tuple{Vararg{Int}}) = x
+normalize(x::Tuple{Vararg{Integer}}) = map(normalize, x)
+normalize(x::Base.OneTo{Int}) = x
+normalize(x::Base.OneTo{<:Integer}) = Base.OneTo{Int}(x)
+normalize(x::AbstractUnitRange{Int}) = x
+normalize(x::AbstractUnitRange{<:Integer}) = AbstractUnitRange{Int}(x)
+normalize(x::AbstractRange{Int}) = x
+normalize(x::AbstractRange{<:Integer}) =
+    normalize(first(x)) : normalize(step(x)) : normalize(last(x))
+normalize(x::CartesianIndex) = x
+normalize(x::CartesianIndices) = x
+@noinline normalize() = error("missing argument in `normalize(x)`")
+@noinline normalize(@nospecialize(x)) =
+    error("unexpected object of type `$(typeof(x))` in `@range` expression, you may specialize `EasyRanges.normalize` for that type")
+
+"""
     EasyRanges.forward(R)
 
 yields an object which contains the same (Cartesian) indices as `R` but with positive
 step(s) and `Int`-valued. Arguments of other types are returned unchanged.
 
 """
-forward(a) = a
-forward(a::AbstractUnitRange{Int}) = a
-forward(a::AbstractUnitRange{<:Integer}) = to_int(a)
-function forward(a::OrdinalRange{<:Integer,<:Integer})
-    first_a, step_a, last_a = first_step_last(a)
-    return step_a ≥ 0 ? (first_a:step_a:last_a) : (last_a:-step_a:first_a)
+forward(x) = _forward(normalize(x))
+
+_forward(i::Int) = i
+_forward(I::CartesianIndex) = I
+_forward(r::AbstractUnitRange{Int}) = r
+_forward(r::OrdinalRange{Int,Int}) = begin
+    first_r, step_r, last_r = first_step_last(r)
+    return step_r ≥ zero(step_r) ?
+        (first_r : step_r : last_r) :
+        (last_r : -step_r : first_r)
 end
-forward(R::CartesianIndices{N,<:NTuple{N,AbstractUnitRange{Int}}}) where {N} = R
-forward(R::CartesianIndices) = CartesianIndices(map(forward, ranges(R)))
+_forward(R::CartesianIndices{N,<:NTuple{N,AbstractUnitRange{Int}}}) where {N} = R
+_forward(R::CartesianIndices) = CartesianIndices(map(forward, ranges(R)))
 
 """
     EasyRanges.backward(R)
@@ -76,147 +125,93 @@ yields an object which constains the same (Cartesian) indices as `R` but with ne
 step(s) and `Int`-valued. Arguments of other types are returned unchanged.
 
 """
-backward(a) = a
-function backward(a::AbstractUnitRange{<:Integer})
-    first_a, last_a = first_last(a)
-    return last_a:-1:first_a
+backward(x) = _backward(normalize(x))
+
+_backward(i::Int) = i
+_backward(I::CartesianIndex) = I
+_backward(r::AbstractUnitRange{Int}) = last(r) : -1 : first(r)
+_backward(r::OrdinalRange{Int,Int}) = begin
+    first_r, step_r, last_r = first_step_last(r)
+    return step_r ≤ zero(step_r) ?
+        (first_r : step_r : last_r) :
+        (last_r : -step_r : first_r)
 end
-function backward(a::OrdinalRange{<:Integer,<:Integer})
-    first_a, step_a, last_a = first_step_last(a)
-    return step_a ≤ 0 ? (first_a:step_a:last_a) : (last_a:-step_a:first_a)
-end
-backward(a::CartesianIndices) = CartesianIndices(map(backward, ranges(a)))
+_backward(R::CartesianIndices) = CartesianIndices(map(backward, ranges(R)))
 
 """
-    EasyRanges.plus(a...)
+    EasyRanges.plus(x)
+    EasyRanges.plus(x, y)
+    EasyRanges.plus(x, y, z)
 
-yields the result of expression `+a`, `a + b`, `a + b + c...` in [`@range`](@ref) macro.
+yield the result of expressions `+x`, `x + y`, or `x + y + z...` in [`@range`](@ref)
+macro.
 
 """ plus
 
-# Use ordinary + by default and deal with multiple arguments.
-plus(a) = +a
-plus(a, b) = a + b
-@inline plus(a, b, c...) = plus(plus(a, b), c...)
+# Unary plus just call `normalize`.
+plus(x) = normalize(x)
 
-# Unary plus just converts to `Int`-valued object.
-plus(a::Int) = a
-plus(a::Integer) = to_int(a)
-plus(a::AbstractUnitRange{Int}) = a
-plus(a::AbstractUnitRange{<:Integer}) = to_int(a)
-plus(a::OrdinalRange{<:Integer,<:Integer}) = forward(a)
-plus(a::CartesianIndex) = a
-plus(a::CartesianIndices) = forward(a)
+# Addition of 2 arguments.
+plus(x, y) = _plus(normalize(x), normalize(y))
 
-# Binary plus.
-plus(a::Integer, b::Integer) = to_int(a) + to_int(b)
-function plus(a::AbstractUnitRange{<:Integer}, b::Integer)
-    first_a, last_a = first_last(a)
-    int_b = to_int(b)
-    return (first_a + int_b):(last_a + int_b)
-end
-plus(a::Integer, b::AbstractUnitRange{<:Integer}) = plus(b, a)
-function plus(a::OrdinalRange{<:Integer,<:Integer}, b::Integer)
-    first_a, step_a, last_a = first_step_last(a)
-    int_b = to_int(b)
-    if step_a ≥ 0
-        return (first_a + int_b):(step_a):(last_a + int_b)
-    else
-        return (last_a + int_b):(-step_a):(first_a + int_b)
-    end
-end
-plus(a::Integer, b::OrdinalRange{<:Integer,<:Integer}) = plus(b, a)
+_plus(x::Int, y::Int) = x + y
+_plus(i::Int, r::AbstractRange{Int}) = _plus(r, i)
+_plus(r::AbstractUnitRange{Int}, i::Int) = first(r) + i : last(r) + i
+_plus(r::OrdinalRange{Int}, i::Int) = first(r) + i : step(r) : last(r) + i
+
+# Addition of 3 or more arguments.
+@inline plus(x, y, z...) = plus(plus(x, y), z...) # FIXME: _plus(normalize(x), plus(y, z...))
 
 """
-    EasyRanges.minus(a...)
+    EasyRanges.minus(x)
+    EasyRanges.minus(x, y)
 
-yields the result of expression `-a` and `a - b` in [`@range`](@ref) macro.
+yield the result of expressions `-x` or `x - y` in [`@range`](@ref) macro.
 
 """ minus
 
-# Use ordinary - by default.
-minus(a) = -a
-minus(a, b) = a - b
+# Unary minus.
+minus(x) = _minus(normalize(x))
 
-# Unary minus yields positive step sign.
-minus(a::Integer) = -to_int(a)
-function minus(a::AbstractUnitRange{<:Integer})
-    first_a, last_a = first_last(a)
-    return (-last_a):(-first_a)
-end
-function minus(a::OrdinalRange{<:Integer,<:Integer})
-    first_a, step_a, last_a = first_step_last(a)
-    if step_a ≥ 0
-        return (-last_a):(step_a):(-first_a)
-    else
-        return (-first_a):(-step_a):(-last_a)
-    end
-end
-minus(a::CartesianIndex) = -a
-minus(a::CartesianIndices) = CartesianIndices(map(minus, ranges(a)))
+_minus(i::Int) = -i
+_minus(r::AbstractUnitRange{Int}) = -last(r) : -first(r)
+_minus(r::OrdinalRange{Int,Int}) = -first(r) : -step(r) : -last(r)
+_minus(I::CartesianIndex) = -I
+_minus(R::CartesianIndices) = CartesianIndices(map(_minus, ranges(R)))
 
-# Binary minus.
-minus(a::Integer, b::Integer) = to_int(a) - to_int(b)
-function minus(a::AbstractUnitRange{<:Integer}, b::Integer)
-    first_a, last_a = first_last(a)
-    int_b = to_int(b)
-    return (first_a - int_b):(last_a - int_b)
-end
-function minus(a::Integer, b::AbstractUnitRange{<:Integer})
-    int_a = to_int(a)
-    first_b, last_b = first_last(b)
-    return (int_a - last_b):(int_a - first_b)
-end
-function minus(a::OrdinalRange{<:Integer,<:Integer}, b::Integer)
-    first_a, step_a, last_a = first_step_last(a)
-    int_b = to_int(b)
-    if step_a ≥ 0
-        return (first_a - int_b):(step_a):(last_a - int_b)
-    else
-        return (last_a - int_b):(-step_a):(first_a - int_b)
-    end
-end
-function minus(a::Integer, b::OrdinalRange{<:Integer,<:Integer})
-    int_a = to_int(a)
-    first_b, step_b, last_b = first_step_last(b)
-    if step_b ≥ 0
-        return (int_a - last_b):(step_b):(int_a - first_b)
-    else
-        return (int_a - first_b):(-step_b):(int_a - last_b)
-    end
-end
+# Subtraction.
+minus(x, y) = _minus(normalize(x), normalize(y))
+
+_minus(x::Int, y::Int) = x - y
+_minus(r::AbstractUnitRange{Int}, i::Int) = first(r) - i : last(r) - i
+_minus(i::Int, r::AbstractUnitRange{Int}) = i - last(r) : i - first(r)
+_minus(r::OrdinalRange{Int,Int}, i::Int) = first(r) - i : step(r) : last(r) - i
+_minus(i::Int, r::OrdinalRange{Int,Int}) = i - first(r) : -step(r) : i - last(r)
 
 """
-    EasyRanges.cap(a...)
+    EasyRanges.cap(a, b)
 
 yields the result of expression `a ∩ b` in [`@range`](@ref) macro.
 
 """
-cap(a, b) = intersect(a, b) # use default behavior
-cap(a::Integer, b::Integer) = cap(to_int(a), to_int(b))
-cap(a::Int, b::Int) = ifelse(a === b, a:a, 1:0)
-cap(a::Integer, b::AbstractUnitRange{<:Integer}) = cap(b, a)
-function cap(a::AbstractUnitRange{<:Integer}, b::Integer)
-    first_a, last_a = first_last(a)
-    int_b = to_int(b)
-    ifelse((first_a ≤ int_b)&(int_b ≤ last_a), int_b:int_b, 1:0)
-end
-cap(a::OneTo, b::OneTo) = OneTo{Int}(min(to_int(a.stop), to_int(b.stop)))
-function cap(a::AbstractUnitRange{<:Integer},
-             b::AbstractUnitRange{<:Integer})
-    first_a, last_a = first_last(a)
-    first_b, last_b = first_last(b)
-    return max(first_a, first_b):min(last_a, last_b)
-end
-function cap(a::OrdinalRange{<:Integer,<:Integer},
-             b::OrdinalRange{<:Integer,<:Integer})
-    return forward(a) ∩ forward(b) # FIXME: Optimize?
-end
-cap(a::CartesianIndex{N}, b::CartesianIndex{N}) where {N} =
-    CartesianIndices(map(cap, Tuple(a), Tuple(b)))
+cap(a, b) = _cap(normalize(a), normalize(b))
 
-# Combine CartesianIndices and CartesianIndices or CartesianIndex.
-for f in (:plus, :minus, :cap)
+_cap(a::Int, b::Int) = ifelse(a == b, a:a, 1:0)
+_cap(i::Int, r::AbstractRange{Int}) = _cap(r, i)
+_cap(r::AbstractUnitRange{Int}, i::Int) = ifelse((first(r) ≤ i)&(i ≤ last(r)), i:i, 1:0)
+_cap(r::AbstractRange{Int}, i::Int) = ifelse(i ∈ r, i:i, 1:0)
+_cap(a::OneTo{Int}, b::OneTo{Int}) = OneTo{Int}(min(a.stop, b.stop))
+_cap(a::AbstractUnitRange{Int}, b::AbstractUnitRange{Int}) =
+    max(first(a), first(b)) : min(last(a), last(b))
+_cap(a::OrdinalRange{Int,Int}, b::OrdinalRange{Int,Int}) =
+    _forward(a) ∩ _forward(b) # FIXME: Optimize?
+_cap(a::CartesianIndex{N}, b::CartesianIndex{N}) where {N} =
+    CartesianIndices(map(_cap, Tuple(a), Tuple(b)))
+_cap(a::CartesianIndices{N}, b::CartesianIndices{N}) where {N} =
+    CartesianIndices(map(_cap, ranges(a), ranges(b)))
+
+# Combine CartesianIndices and a CartesianIndex.
+for f in (:_plus, :_minus, :_cap)
     @eval begin
         $f(a::CartesianIndices{N}, b::CartesianIndex{N}) where {N} =
             CartesianIndices(map($f, ranges(a), Tuple(b)))
@@ -225,9 +220,6 @@ for f in (:plus, :minus, :cap)
     end
 end
 
-cap(a::CartesianIndices{N}, b::CartesianIndices{N}) where {N} =
-    CartesianIndices(map(cap, ranges(a), ranges(b)))
-
 """
     EasyRanges.stretch(a, b)
 
@@ -235,21 +227,14 @@ yields the result of stretching `a` by amount `b`. This is equivalent to the exp
 ± b` in [`@range`](@ref) macro.
 
 """
-stretch(a::Int, b::Int) = (a - b):(a + b)
-function stretch(a::AbstractUnitRange{<:Integer}, b::Integer)
-    first_a, last_a = first_last(a)
-    int_b = to_int(b)
-    return (first_a - int_b):(last_a + int_b)
-end
-function stretch(a::OrdinalRange{<:Integer}, b::Integer)
-    first_a, step_a, last_a = first_step_last(a)
-    int_b = to_int(b)
-    (int_b % step_a) == 0 || throw(ArgumentError("stretch must be multiple of the step"))
-    if step_a ≥ 0
-        return (first_a - int_b):step_a:(last_a + int_b)
-    else
-        return (last_a - int_b):(-step_a):(first_a + int_b)
-    end
+stretch(a, b) = _stretch(normalize(a), normalize(b))
+
+_stretch(a::Int, b::Int) = a - b : a + b
+_stretch(r::AbstractUnitRange{Int}, i::Int) = first(r) - i : last(r) + i
+_stretch(r::OrdinalRange{Int,Int}, i::Int) = begin
+    s = step(r)
+    iszero(i % s) || throw(ArgumentError("stretch must be multiple of the step"))
+    return first(r) - i : s : last(r) + i
 end
 
 """
@@ -259,43 +244,34 @@ yields the result of shrinking `a` by amount `b`. This is equivalent to the expr
 ∓ b` in [`@range`](@ref) macro.
 
 """
-shrink(a::Int, b::Int) = (a + b):(a - b)
-function shrink(a::AbstractUnitRange{<:Integer}, b::Integer)
-    first_a, last_a = first_last(a)
-    int_b = to_int(b)
-    return (first_a + int_b):(last_a - int_b)
-end
-function shrink(a::OrdinalRange{<:Integer}, b::Integer)
-    first_a, step_a, last_a = first_step_last(a)
-    int_b = to_int(b)
-    (int_b % step_a) == 0 || throw(ArgumentError("shrink must be multiple of the step"))
-    if step_a ≥ 0
-        return (first_a + int_b):step_a:(last_a - int_b)
-    else
-        return (last_a + int_b):(-step_a):(first_a - int_b)
-    end
+shrink(a, b) = _shrink(normalize(a), normalize(b))
+
+_shrink(a::Int, b::Int) = a + b : a - b
+_shrink(r::AbstractUnitRange{Int}, i::Int) = first(r) + i : last(r) - i
+_shrink(r::OrdinalRange{Int,Int}, i::Int) = begin
+    s = step(r)
+    iszero(i % s) || throw(ArgumentError("shrink must be multiple of the step"))
+    return first(r) + i : s : last(r) - i
 end
 
-for f in (:stretch, :shrink)
+for f in (:_stretch, :_shrink)
     @eval begin
-        $f(a::Integer, b::Integer) = $f(to_int(a), to_int(b))
-
-        $f(a::CartesianIndices{N}, b::CartesianIndex{N}) where {N} =
-            CartesianIndices(map($f, ranges(a), Tuple(b)))
-        $f(a::CartesianIndices{N}, b::NTuple{N,Integer}) where {N} =
-            CartesianIndices(map($f, ranges(a), b))
-        $f(a::CartesianIndices, b::Integer) =
-            CartesianIndices(map(Base.Fix2($f, b), ranges(a)))
+        $f(R::CartesianIndices{N}, I::CartesianIndex{N}) where {N} =
+            CartesianIndices(map($f, ranges(R), Tuple(I)))
+        $f(R::CartesianIndices{N}, I::NTuple{N,Int}) where {N} =
+            CartesianIndices(map($f, ranges(R), I))
+        $f(R::CartesianIndices, i::Int) =
+            CartesianIndices(map(Base.Fix2($f, i), ranges(R)))
     end
     # A Cartesian index can be stretched, not shrinked.
-    if f === :stretch
+    if f === :_stretch
         @eval begin
             $f(a::CartesianIndex{N}, b::CartesianIndex{N}) where {N} =
                 CartesianIndices(map($f, Tuple(a), Tuple(b)))
             $f(a::CartesianIndex{N}, b::NTuple{N,Integer}) where {N} =
                 CartesianIndices(map($f, Tuple(a), b))
-            $f(a::CartesianIndex, b::Integer) =
-                CartesianIndices(map(Base.Fix2($f, b), Tuple(a)))
+            $f(I::CartesianIndex, i::Int) =
+                CartesianIndices(map(Base.Fix2($f, i), Tuple(I)))
         end
     end
 end
@@ -315,7 +291,7 @@ yields the 2-tuple `(first(x), last(x))` converted to be `Int`-valued.
 
 """
 first_last(x::AbstractUnitRange{<:Integer}) =
-    (to_int(first(x)), to_int(last(x)))
+    (normalize(first(x)), normalize(last(x)))
 
 first_last(x::CartesianIndices) = begin
     flag = true
@@ -334,10 +310,10 @@ yields the 3-tuple `(first(x), step(x), last(x))` converted to be `Int`-valued.
 
 """
 first_step_last(x::AbstractUnitRange{<:Integer}) =
-    (to_int(first(x)), 1, to_int(last(x)))
+    (normalize(first(x)), 1, normalize(last(x)))
 
 first_step_last(x::OrdinalRange{<:Integer,<:Integer}) =
-    (to_int(first(x)), to_int(step(x)), to_int(last(x)))
+    (normalize(first(x)), normalize(step(x)), normalize(last(x)))
 
 first_step_last(x::CartesianIndices) =
     (CartesianIndex(map(first, ranges(x))),